Convert X,Y,Z coordinates of ROIs into integer numbers before further use.

+++ fileops/export/_roi.py
@@ -23,10 +23,10 @@ class ConfigROI(NamedTuple):
 
 
 def rectangle_roi(rect_p: rect_params, center_is_middle=True) -> ImagejRoi:
-    x0 = (rect_p.X - rect_p.W / 2) if center_is_middle else rect_p.X
-    y0 = (rect_p.Y - rect_p.H / 2) if center_is_middle else rect_p.Y
-    x1 = (rect_p.X + rect_p.W / 2) if center_is_middle else rect_p.X + rect_p.W
-    y1 = (rect_p.Y + rect_p.H / 2) if center_is_middle else rect_p.Y + rect_p.H
+    x0 = int(rect_p.X - rect_p.W / 2) if center_is_middle else rect_p.X
+    y0 = int(rect_p.Y - rect_p.H / 2) if center_is_middle else rect_p.Y
+    x1 = int(rect_p.X + rect_p.W / 2) if center_is_middle else rect_p.X + rect_p.W
+    y1 = int(rect_p.Y + rect_p.H / 2) if center_is_middle else rect_p.Y + rect_p.H
     rect_roi = ImagejRoi.frompoints(np.array([[x0, y0], [x1, y0], [x1, y1], [x0, y1]]))
     rect_roi.roitype = ROI_TYPE.RECT
